pyproject.toml
[project]
name = "agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40",
    "openai>=1.50",
    "google-genai>=1.0",
    "python-telegram-bot>=21",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "aiohttp>=3.10",
    "aiosqlite>=0.20",
    "structlog>=24.0",
    "discord-py>=2.4",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.ruff]
line-length = 100
